reorganizing App and Homepage
