+++ client/src/components/App.jsx
@@ -1,168 +1,35 @@
+/**
+ *
+ * App.jsx
+ *
+ * This component is the skeleton around the actual pages, and should only
+ * contain code that should be seen on all pages. (e.g. navigation bar)
+ */
+
+
+
 import React, { Component } from 'react';
-import Sidebar from 'react-sidebar';
-import MentorList from '../containers/mentor_list.jsx';
-import { filter } from 'lodash';
-
-const mql = window.matchMedia(`(min-width: 800px)`);
-
-const techStackItems = [
-  'All',
-  'React',
-  'Angular',
-  'Backbone',
-  'React Native',
-  'Express',
-  'Node.js',
-  'TDD',
-  'Mocha/Chai',
-  'Redux'
-];
-
-const rolesItems = [
-  'All',
-  'Full Stack',
-  'Front-end',
-  'Back-end'
-];
-
-const locationItems = [
-  'All',
-  'San Francisco',
-  'San Jose',
-  'Palo Alto'
-];
-
+import Nav from './Nav.react';
+import { connect } from 'react-redux';
+// import auth from '../utils/auth';
 class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      mql: mql,
-      docked: props.docked,
-      open: props.open,
-      isChecked: false,
-      selectedTechStacksItems: [],
-      selectedRolesItems: [],
-      selectedLocationItems: []
-    }
-    this.mediaQueryChanged = this.mediaQueryChanged.bind(this);
-    this.onSetSidebarOpen = this.onSetSidebarOpen.bind(this);
-    this.createTechStackCheckboxes = this.createTechStackCheckboxes.bind(this);
-    this.handleInputChange = this.handleInputChange.bind(this);
-  }
-
-  onSetSidebarOpen(open) {
-    this.setState({sidebarOpen: open});
-  }
-
-  componentWillMount() {
-    mql.addListener(this.mediaQueryChanged);
-    this.setState({mql: mql, sidebarDocked: mql.matches});
-  }
-
-  mediaQueryChanged() {
-    this.setState({sidebarDocked: this.state.mql.matches});
-  }
-
-
-  handleInputChange(event) {
-    const target = event.target;
-    this.setState({
-      isChecked: target
-    });
-  }
-
-  createTechStackCheckboxes() {
-    return (
-      <form>
-        {techStackItems.map((item, key) => {
-          return (
-            <div key={key}>
-              <input 
-                key={key}
-                id={key}
-                type="checkbox"
-                checked={this.state.isChecked[key]}
-                onChange={this.handleInputChange} />
-              <label htmlFor={key}>{item}</label>
-            </div>
-          )
-        })}
-      </form>
-    );
-  }
-
-  createRolesCheckboxes() {
-    return (
-      <form>
-        {rolesItems.map((item, key) => {
-          return (
-            <div key={key}>
-              <input 
-                key={key}
-                id={key}
-                type="checkbox"
-                checked={this.state.isChecked[key]}
-                onChange={this.handleInputChange} />
-              <label htmlFor={key}>{item}</label>
-            </div>
-          )
-        })}
-      </form>
-    );
-  }
-
-  createLocationCheckboxes() {
-    return (
-      <form>
-        {locationItems.map((item, key) => {
-          return (
-            <div key={key}>
-              <input 
-                key={key}
-                id={key}
-                type="checkbox"
-                checked={this.state.isChecked[key]}
-                onChange={this.handleInputChange} />
-              <label htmlFor={key}>{item}</label>
-            </div>
-          )
-        })}
-      </form>
-    );
-  }
-
   render() {
-    var sidebarContent = 
-    (<div>
-      <div>Tech Stacks</div>
-      <div>
-        {this.createTechStackCheckboxes()}
-      </div>
-      <br />
-      <div>Roles</div>
-      <div>
-        {this.createRolesCheckboxes()}
-      </div>
-      <br />
-      <div>Location</div>
-      <div>
-        {this.createLocationCheckboxes()}
-      </div>
-    </div>);
-
-
     return (
-      <div>
-        <Sidebar 
-          sidebar={sidebarContent}
-          open={this.state.sidebarOpen}
-          docked={this.state.sidebarDocked}
-          onSetOpen={this.onSetSidebarOpen}>
-          <MentorList />
-        </Sidebar>
+       <div>
+        <Nav loggedIn={this.props.data.loggedIn} history={this.props.history} location={this.props.location} dispatch={this.props.dispatch} currentlySending={this.props.data.currentlySending} />
+        { this.props.children }
       </div>
-    );
+    )
   }
-};
+}
 
-export default App;
+// export default App
+function select(state) {
+  return {
+    data: state
+  };
+}
+
+// Wrap the component to inject dispatch and state into it
+export default connect(select)(App);
+
